asdsad

+++ table/hash_table/src/data.h
@@ -4,7 +4,7 @@
 #include <stdint.h>
 
 typedef struct entry entry_t;
-typedef uint32_t table_val_t;
+typedef uint64_t table_val_t;
 typedef entry_t list_data_t;
 
 struct entry
