+++ table/hash_table/src/hash_table.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstdint>
 #include <immintrin.h>
 #include <stdbool.h>
 #include <stdint.h>
@@ -13,11 +14,11 @@
 typedef struct hash_table
 {
 	list_t *buckets;
-	uint32_t size;
+	uint64_t size;
 	// uint32_t (*hash_function)(__m256i);
 } hash_table_t;
 
-hash_table_t table_init(uint32_t sz);
+hash_table_t table_init(uint64_t sz);
 hash_table_t build_table_from_text(char *text);
 void table_free(hash_table_t *table);
 table_val_t *table_get_key(hash_table_t *table, char *key, size_t len);
